Push - Add a top-level 'push' command to the CLI
Split the functionality of the outbox workflow.
Instead of copying the files to the outbox and pushing to the central
store, just focus on copying the files.
The push command will push the packages to the central store.

The split should model this to be more de-centralized.  Will allow users
to choose when they want to deploy what they packaged.